fix(api): change Increment field type from int to int64 in Metric struct

+++ src/server/api/go/internal/modules/model/metric.go
@@ -11,7 +11,7 @@ type Metric struct {
 	ProjectID uuid.UUID `gorm:"type:uuid;not null;index:idx_metric_project_id_tag_created_at,priority:1" json:"project_id"`
 
 	Tag       string `gorm:"type:text;not null;index:idx_metric_project_id_tag_created_at,priority:2" json:"tag"`
-	Increment int    `gorm:"not null;default:0" json:"increment"`
+	Increment int64  `gorm:"not null;default:0" json:"increment"`
 
 	CreatedAt time.Time `gorm:"autoCreateTime;not null;default:CURRENT_TIMESTAMP;index:idx_metric_project_id_tag_created_at,priority:3" json:"created_at"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime;not null;default:CURRENT_TIMESTAMP" json:"updated_at"`
